Add include file to forward-declare inline assembly structs.
