+++ libsolidity/inlineasm/AsmAnalysis.h
@@ -26,6 +26,8 @@
 
 #include <libjulia/backends/evm/AbstractAssembly.h>
 
+#include <libsolidity/inlineasm/AsmDataForward.h>
+
 #include <boost/variant.hpp>
 
 #include <functional>
@@ -38,21 +40,6 @@ namespace solidity
 class ErrorReporter;
 namespace assembly
 {
-
-struct Literal;
-struct Block;
-struct Label;
-struct FunctionalInstruction;
-struct Assignment;
-struct VariableDeclaration;
-struct Instruction;
-struct Identifier;
-struct StackAssignment;
-struct FunctionDefinition;
-struct FunctionCall;
-struct Switch;
-struct ForLoop;
-using Statement = boost::variant<Instruction, Literal, Label, StackAssignment, Identifier, Assignment, FunctionCall, FunctionalInstruction, VariableDeclaration, FunctionDefinition, Switch, Block, ForLoop>;
 
 struct AsmAnalysisInfo;
